refactor: move `rows` TypeORM object to JS side
